fix(fetch jaen data): cope with the caching issue by removing dept to rawgit
The fetching does no longer depend on raw.githubusercontent.co. This has been introduced in order to
cope with a caching issue where the jaen-data.json has been cached for five minutes. In addition the
fetch function is no longer located in the cmsProvider but in a new action.

ref: https://github.com/rgrove/rawgit/blob/master/FAQ.md#cdn-ttl
